test: move normalize options test to seperate file
