+++ test/ports-map.js
@@ -77,6 +77,7 @@ const listOfTests = {
   "client-reconnect-option": 1,
   "cli-server": 1,
   "server-option": 1,
+  "normalize-option": 1,
 };
 
 let startPort = 8089;
